@@ -23,7 +23,7 @@
     "autoprefixer": "^10.4.0",
     "axios": "^1.6.0",
     "lucide-react": "^0.400.0",
-    "next": "15.0.0",
+    "next": "^15.5.6",
     "postcss": "^8.4.0",
     "react": "^18.3.0",
     "react-dom": "^18.3.0",
