@@ -7,7 +7,7 @@ from channels.db import database_sync_to_async
 from rest_framework_simplejwt.tokens import UntypedToken
 from rest_framework_simplejwt.exceptions import InvalidToken, TokenError
 from django.contrib.auth import get_user_model
-from django.conf import settings
+# Note: settings import removed - not used in this file
 from apps.chat.models import ChatMessage, TicketParticipant
 
 
