@@ -4,9 +4,10 @@ Notification model for Communication Service.
 import sys
 from pathlib import Path
 
-# Add shared directory to path
-project_root = Path(__file__).resolve().parent.parent.parent.parent.parent
-shared_path = project_root / 'shared' / 'core'
+# Add shared directory to path (try Docker mount path first)
+docker_shared_path = Path('/shared/core')
+local_shared_path = Path(__file__).resolve().parent.parent.parent.parent.parent / 'shared' / 'core'
+shared_path = docker_shared_path if docker_shared_path.exists() else local_shared_path
 if str(shared_path) not in sys.path:
     sys.path.insert(0, str(shared_path))
 
