@@ -57,7 +57,7 @@ const AssigneeReportsPage: React.FC = () => {
     const fetchTickets = async () => {
       try {
         const response = await ticketService.getTickets({ assigneeId: user?.id });
-        const ticketsList = Array.isArray(response) ? response : (response?.results || []);
+        const ticketsList = response.results || [];
         setTickets(ticketsList);
       } catch (error: any) {
         console.error('Error fetching tickets:', error);
