@@ -51,7 +51,7 @@ const MyTasksPage: React.FC = () => {
     const fetchTickets = async () => {
       try {
         const response = await ticketService.getTickets({ assigneeId: user?.id });
-        const ticketsList = Array.isArray(response) ? response : (response?.results || []);
+        const ticketsList = response.results || [];
         setTickets(ticketsList);
       } catch (error: any) {
         console.error('Error fetching tickets:', error);
@@ -263,7 +263,7 @@ const MyTasksPage: React.FC = () => {
             }
             // Refresh tickets
             const response = await ticketService.getTickets({ assigneeId: user?.id });
-            const ticketsList = Array.isArray(response) ? response : (response?.results || []);
+            const ticketsList = response.results || [];
             setTickets(ticketsList);
           } finally {
             setProcessing(false);
