@@ -26,7 +26,7 @@ const AdminDashboard: React.FC = () => {
     const fetchTickets = async () => {
       try {
         const response = await ticketService.getTickets();
-        const ticketsList = Array.isArray(response) ? response : (response?.results || []);
+        const ticketsList = response.results || [];
         setTickets(ticketsList);
       } catch (error: any) {
         // Handle network errors gracefully - API might not be available
