@@ -41,7 +41,7 @@ const AssigneeDashboard: React.FC = () => {
     const fetchTickets = async () => {
       try {
         const response = await ticketService.getTickets({ assigneeId: user?.id });
-        const ticketsList = Array.isArray(response) ? response : (response?.results || []);
+        const ticketsList = response.results || [];
         setTickets(ticketsList);
       } catch (error: any) {
         // Handle network errors gracefully - API might not be available
